nas_agent_graph/requirements.txt
# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Pydantic for structured output
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Phoenix observability
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0
openinference-instrumentation-openai>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Existing nas_system dependencies (inherited)
# torch, numpy, pandas, matplotlib, scikit-learn, PyYAML
